new program2 for pitch/vol variation

+++ software/apps/program2/final_design/hr_sc04.c
@@ -30,8 +30,8 @@ void timer_event_handler(nrf_timer_event_t event_type, void* p_context) {
 }
 
 //measure dist with trig, echo, and timer_channel passed in
- uint32_t hr_sc04_measure_dist(uint32_t trig, uint32_t echo, nrf_timer_cc_channel_t timer_channel){
-    uint32_t distance;
+ float hr_sc04_measure_dist(uint32_t trig, uint32_t echo, nrf_timer_cc_channel_t timer_channel){
+    float distance;
     uint32_t duration = 0;
 
     nrf_gpio_pin_write(trig, 1);
